Reconfig classCards;Extract form from newClassModal>child component
